+++ python/samples/single_live_sensor.py
@@ -3,7 +3,6 @@
 import pprint
 
 import cepton_sdk
-
 import common
 
 if __name__ == "__main__":